First Commit
Three unity projects have been uploded.
1. Chain Reaction:
The initial version of the original Android game.
2. Chain Recation - Improvised:
A rotation factor is added to the spheres; The collision script and animation are yet to be complete.
3. Chain Reaction - Spatial - 1:
The Game is rethought using spatial forms. Spheres are replaced with rooms and subdivision is shown, A score generator has been added. The unity recorder has also been integrated to record Game Play